Module gdf_translate::google::gcloud::translate::v3

implementation of google translation api v3

See following links

Google Translate V3 Intro

translateText API

batchTranslateText API

batch translation result -long running operation

get long running operatopm result - short polling approach

get long running operatopm result - long polling approach

Sample curls:

Initiate batch translation

curl --location --request POST 'https://translation.googleapis.com/v3/projects/dummy-project-id/locations/us-central1:batchTranslateText' \
--header 'Authorization: Bearer ya29.c....' \
--header 'Content-Type: application/javascript' \
--data-raw '{
    "sourceLanguageCode": "en",
    "targetLanguageCodes": "de",
    "inputConfigs": [{
        "mimeType":  "text/html",
        "gcsSource": {
            "inputUri": "gs://translate_v3_test_in/input.tsv"
        }
    }],
    "outputConfig": {
        "gcsDestination": {
            "outputUriPrefix": "gs://translate_v3_test_out/"
        }
    }
}'

Check long running operation status with immediate response (kind of short polling)

curl --location --request GET 'https://translation.googleapis.com/v3/projects/dummy-project-id/locations/us-central1/operations/20200615-11581592247524-5edeccd9-0000-26b7-bd4f-30fd38139c64' \
--header 'Authorization: Bearer ya29.c....' \

Check long running operation status with delayed response (kind of long polling)

curl --location --request POST 'https://translation.googleapis.com/v3/projects/dummy-project-id/locations/us-central1/operations/20200615-11581592247524-5edeccd9-0000-26b7-bd4f-30fd38139c64:wait' \
--header 'Authorization: Bearer ya29.c....' \
--header 'Content-Type: application/json' \
--data-raw '{
  "timeout": "60s"
}'

Structs

GoogleTranslateV3ResponseMetadata

GoogleTranslateV3Response

GoogleTranslateV3ApiResponse

GoogleTranslateV3WaitResponseMetadata

GoogleTranslateV3WaitResponseMeResponse

GoogleTranslateV3WaitResponseErrorDetail

GoogleTranslateV3WaitResponseError

GoogleTranslateV3WaitResponse

GoogleTranslateV3WaitApiResponse

Functions

batch_translate_text
Translates csv/tsv file using Google Translate V3 REST API

batch_translate_text_check_status
Check the status of long running operation representing batch translation request

map_to_string

string_to_map
Converts string produced by Google Translate V3 API back to translation map